// src/api/generated/models/pageResultAppCaseSummary.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AppCaseSummary } from './appCaseSummary';

export interface PageResultAppCaseSummary {
  list?: AppCaseSummary[];
  total?: string;
  pageNum?: number;
  pageSize?: number;
  totalPages?: number;
  hasNext?: boolean;
  hasPrevious?: boolean;
}